Extract TraefikParser to its own file
